+++ src/lib/chimp-helper.js
@@ -14,6 +14,11 @@ var chai            = require('chai'),
     fs              = require('fs-extra'),
     exit            = require('exit'),
     booleanHelper   = require('./boolean-helper');
+import merge from 'deepmerge';
+import {
+  parseNullableString,
+  parseNullableInteger,
+  parseBoolean } from './environment-variable-parsers';
 
 var chimpHelper = {
   loadAssertionLibrary: function () {
@@ -65,76 +70,44 @@ var chimpHelper = {
 
     var setupBrowser = function () {
       log.debug('[chimp][helper] getting browser');
-      var customChimpConfigPath = path.resolve(process.cwd(), process.env['chimp.path'], 'chimp.js');
 
-      var _translateLogLevel = function () {
-        if (booleanHelper.isTruthy(process.env['chimp.webdriverLogLevel'])) {
-          return process.env['chimp.webdriverLogLevel'];
-        } else if (process.env['chimp.log'] === 'info' ||
-          process.env['chimp.log'] === 'warn' ||
-          process.env['chimp.log'] === 'error') {
-          return 'silent'
-        } else {
-          return process.env['chimp.log'];
-        }
-      };
-
-      global.sessionManager = new SessionFactory({
-        host: process.env['chimp.host'],
-        port: process.env['chimp.port'],
-        user: process.env['chimp.user'],
-        key: process.env['chimp.key'],
-        browser: process.env['chimp.browser'],
-        deviceName: process.env['chimp.deviceName']
-      });
-
-      if (fs.existsSync(customChimpConfigPath)) {
-        var customChimpConfigurator = wrapAsync(require(customChimpConfigPath));
-        global.browser = customChimpConfigurator(global.sessionManager);
-      } else {
-        log.debug('[chimp][helper] custom chimp.js not found, loading defaults');
-        var webdriverOptions = {
-          waitforTimeout: parseInt(process.env['chimp.waitForTimeout']),
-          timeoutsImplicitWait: parseInt(process.env['chimp.timeoutsImplicitWait']),
+      const webdriverioConfigOptions = JSON.parse(process.env['chimp.webdriverio']);
+      const webdriverioOptions = merge(
+        webdriverioConfigOptions,
+        {
           desiredCapabilities: {
-            browserName: process.env['chimp.browser'],
-            platform: process.env['chimp.platform'],
-            name: process.env['chimp.name'],
-            version: process.env['chimp.version']
+            browserName: parseNullableString(process.env['chimp.browser']),
+            platform: parseNullableString(process.env['chimp.platform']),
+            name: parseNullableString(process.env['chimp.name']),
+            version: parseNullableString(process.env['chimp.version']),
+            deviceName: parseNullableString(process.env['chimp.deviceName']),
           },
-          user: process.env['chimp.user'] || process.env.SAUCE_USERNAME,
-          key: process.env['chimp.key'] || process.env.SAUCE_ACCESS_KEY,
-          host: process.env['chimp.host'],
-          port: process.env['chimp.port'],
-          logLevel: _translateLogLevel(),
-          screenshotPath: process.env['chimp.screenshotPath'],
-          sync: booleanHelper.isTruthy(process.env['chimp.sync'])
-        };
-
-        webdriverOptions.desiredCapabilities.chromeOptions = webdriverOptions.desiredCapabilities.chromeOptions || {};
-        if (booleanHelper.isTruthy(process.env['chimp.chromeBin'])) {
-          webdriverOptions.desiredCapabilities.chromeOptions.binary = process.env['chimp.chromeBin'];
+          user: parseNullableString(process.env['chimp.user'] || process.env.SAUCE_USERNAME),
+          key: parseNullableString(process.env['chimp.key'] || process.env.SAUCE_ACCESS_KEY),
+          host: parseNullableString(process.env['chimp.host']),
+          port: parseNullableInteger(process.env['chimp.port']),
+          logLevel: booleanHelper.isTruthy(process.env['chimp.debug']) ?
+            'verbose' : webdriverioConfigOptions.logLevel,
+          sync: parseBoolean(process.env['chimp.sync']),
         }
-        if (booleanHelper.isTruthy(process.env['chimp.chromeArgs'])) {
-          webdriverOptions.desiredCapabilities.chromeOptions.args = process.env['chimp.chromeArgs'].split(',');
-        } else if (booleanHelper.isTruthy(process.env['chimp.chromeNoSandbox'])) {
-          webdriverOptions.desiredCapabilities.chromeOptions.args = ['no-sandbox'];
-        }
+      );
 
-        if (booleanHelper.isTruthy(process.env['chimp.baseUrl'])) {
-          webdriverOptions.baseUrl = process.env['chimp.baseUrl']
+      global.sessionManager = new SessionFactory(Object.assign(
+        _.pick(webdriverioOptions, 'host', 'port', 'user', 'key'),
+        {
+          browser: webdriverioOptions.desiredCapabilities.browserName,
+          deviceName: webdriverioOptions.desiredCapabilities.deviceName,
         }
-        if (process.env['chimp.watch']) {
-          webdriverOptions.desiredCapabilities.applicationCacheEnabled = false;
-        }
-        webdriverOptions.desiredCapabilities['tunnelIdentifier'] = process.env['chimp.tunnelIdentifier'];
-        webdriverOptions.desiredCapabilities['browserstack.local'] = process.env['chimp.browserstackLocal'];
+      ));
 
-        log.debug('[chimp][helper] webdriverOptions are ', JSON.stringify(webdriverOptions));
-
-        var remoteSession = wrapAsync(global.sessionManager.remote, global.sessionManager);
-        global.browser = remoteSession(webdriverOptions);
+      if (booleanHelper.isTruthy(process.env['chimp.watch'])) {
+        webdriverioOptions.desiredCapabilities.applicationCacheEnabled = false;
       }
+
+      log.debug('[chimp][helper] webdriverioOptions are ', JSON.stringify(webdriverioOptions));
+
+      const remoteSession = wrapAsync(global.sessionManager.remote, global.sessionManager);
+      global.browser = remoteSession(webdriverioOptions);
 
       chaiAsPromised.transferPromiseness = global.browser.transferPromiseness;
     };
@@ -155,9 +128,6 @@ var chimpHelper = {
         this.saveScreenshot(ssPath, false);
         log.debug('[chimp][helper] saved screenshot to', ssPath);
       });
-
-      browser.timeoutsAsyncScriptSync(parseInt(process.env['chimp.timeoutsAsyncScript']));
-      log.debug('[chimp][helper] set timeoutsAsyncScript');
 
       if (process.env['chimp.browser'] === 'phantomjs') {
         browser.setViewportSizeSync({